Fixed following issues
In Leads module , Email field is not getting displayed. closes #89
Text "Please wait loading data" is not displayed in Center, when the device is in Land scape mode. closes #91
In opportunities module , Field "Account name" is not displayed. closes #96
In campaigns module , Start date and end date are not displayed. closes #98